+++ sdk/js/src/token_bridge/getIsWrappedAsset.ts
@@ -1,7 +1,7 @@
 import { Connection, PublicKey } from "@solana/web3.js";
+import { LCDClient } from "@terra-money/terra.js";
 import { ethers } from "ethers";
 import { Bridge__factory } from "../ethers-contracts";
-import { ConnectedWallet as TerraConnectedWallet } from "@terra-money/wallet-provider";
 import { importTokenWasm } from "../solana/wasm";
 
 /**
@@ -23,7 +23,7 @@ export async function getIsWrappedAssetEth(
 
 export async function getIsWrappedAssetTerra(
   tokenBridgeAddress: string,
-  wallet: TerraConnectedWallet,
+  client: LCDClient,
   assetAddress: string
 ) {
   return false;