+++ sdk/js/src/utils/consts.ts
@@ -14,3 +14,6 @@ export const CHAIN_ID_ETHEREUM_ROPSTEN: ChainId = 10001;
 export const WSOL_ADDRESS = "So11111111111111111111111111111111111111112";
 export const WSOL_DECIMALS = 9;
 export const MAX_VAA_DECIMALS = 8;
+
+export const TERRA_REDEEMED_CHECK_WALLET_ADDRESS =
+  "terra1x46rqay4d3cssq8gxxvqz8xt6nwlz4td20k38v";