+++ sdk/js/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.2.0
+
+### Changed
+
+Updated @terra-money/terra.js to 3.0.7
+
+Removed @terra-money/wallet-provider
+
+Removed walletAddress parameter from getIsTransferCompletedTerra
+
 ## 0.1.7
 
 ### Added